more checks added for board data

+++ main.js
@@ -3,7 +3,7 @@ import { row, rowMixed1, rowMixed2, rowTriple } from './js/moduls/rowMaker.js';
 import toggleMenu from './js/moduls/headerAnim.js';
 import { render } from '../node_modules/lit-html/lit-html.js';
 import { footerTemplate } from './js/moduls/renderFunctions.js';
-import {today, time} from './js/moduls/clock.js';
+import { today, time } from './js/moduls/clock.js';
 
 toggleMenu();
 
@@ -23,8 +23,8 @@ function start(event) {
   document.querySelector('.minNumInput').setAttribute('disabled', '');
   document.querySelector('.maxNumInput').setAttribute('disabled', '');
   document.querySelector('.controlInput').setAttribute('disabled', '');
-  document.querySelector('.joker50').setAttribute('disabled', '');
-  document.querySelector('.joker100').setAttribute('disabled', '');
+  document.querySelector('.hint50').setAttribute('disabled', '');
+  document.querySelector('.hint100').setAttribute('disabled', '');
   eStartButton.style.setProperty('display', 'none');
   document.querySelector('.submit').style.setProperty('display', 'block');
   setInterval(incrementSeconds, 1000);
@@ -167,17 +167,27 @@ function submit() {
         .querySelector('ul :nth-child(6) a')
         .setAttribute('href', './html/comics.html');
     }
-    
+
     const minNumberInput = document.querySelector('.minNumInput').value;
     const maxNumberInput = document.querySelector('.maxNumInput').value;
     const inputRowsNumber = document.querySelector('.controlInput').value;
-    if (minNumberInput == 0 && maxNumberInput == 10 && inputRowsNumber == 2) {
+    const hint50 = document.querySelector('.hint50').checked;
+    const hint100 = document.querySelector('.hint100').checked;
+    if (
+      minNumberInput == 0 &&
+      maxNumberInput == 10 &&
+      inputRowsNumber == 2 &&
+      hint50 != true &&
+      hint100 != true
+    ) {
+      if (condition) {
+      }
       let newRecord = {
-        date: today.toLocaleDateString('en-GB'), 
+        date: today.toLocaleDateString('en-GB'),
         time: time,
-        record: seconds
+        record: seconds,
       };
-      let recordsArr = JSON.parse(localStorage.recordsArr); 
+      let recordsArr = JSON.parse(localStorage.recordsArr);
       recordsArr.push(newRecord);
       localStorage.clear();
       localStorage.recordsArr = JSON.stringify(recordsArr);
